fix: Fix regression in previous patch
We need to adjust the beginning of the range after reading a value since
we're no longer incrementing the node key index (due to an issue with
the eBPF verifier).

If we don't adjust the beginning of the range we keep grabbing the same
value over and over in an infinite loop.